Fixed warnings are errors when checking with Dymola.
Classes extending Layer are blocks; to extend a model in a block is invalid Modelica => changed Layer type from model to block.

The example networks connect in- and outputs with layers via component modification. This in invalid Modelica => changed to proper connect equation.

The NARX_Network example leverages on propagating sizes via the connector. Again, this is not supported by the Modelica specification. => Sizes are now provided via redeclaration of the in- and outputs with size information. General comment: The current library design breaks with MSL design, where MIMO interfaces must have size parameters that are used to properly set the sizes. This is a rather big design issue.